refactor: separate auth and user context
